requirements.txt
customtkinter
pymem
pillow
psutil